Add targets to build binaries
Currently the Makefile supports building a binary for the host system
and a binary for linux on AMD64. In the future we may add targets for
other OSes and architectures, e.g. linux on ARM.

Additionally we have the binary targets depend on all go files to ensure
the binaries are rebuilt as soon as a go file changes. Since this is
handy for the .coverage.out file as well, we added a depency there too.