Refactor the httpapi package
It does not contain an HTTP server anymore, instead it creates an
http.Handler capable of routing the various endpoints of the public,
non-encrypted HTTP API.

The HTTP server has been moved to the server package. It is not exported
anymore. Instead a Server type has been created which takes care of
starting acmeproxy's various protocol servers.