Make acmeproxy obtain certificates for itself
After setting this aside for a few days I changed my mind about the
Agent. The initial idea as reflected by the comment I wrote back then
was way to complicated.

The Agent.CreateAccount method enables callers to create an ACME account
associated with an unique clientID. This clientID is then used when
calling Agent.RegisterDomain to associate a domain with a client of
acmeproxy. RegisterDomain immediately obtains certificates for the
registered domains.

The acmeproxy server calls those methods during startup in order to
obtain certificates for itself.

During implementation I added a lot of TODOs which will be resolved as
soon as I stabilized acmeproxy's initial architecture.