Refactor reads and updates bbolt buckets
In order to improve the tests of readRecord it became obvious that
its signature has to change. This eventually led to the introduction of
a bucket type which now provides readRecord and an analogous writeRecord
function. The bucket keeps track of errors that may have occured during
its lifetime and does not execute further actions if it has an error.
This allows to simplify the error handling in the various users of
bucket.