Use 10.30.50.1 as the default IP for challtestsrv
The address 10.30.50.1 is the default gateway address of the internal
docker network. At least on my linux laptop it seems to be possible to
access services on the host machine when using it.

Unfortunately using host.docker.internal does not seem to work. If this
does not work reliably across Linux and Mac I will have to think of
something else.