Restructure api.Server start-up
The api.Server needs to start multiple components in separate Go
routines. Additionally it needs to obtain a TLS certificate for itself
from Let's Encrypt. The previous start-up implementation did not take
this into account. The new structure should do the trick, but is not yet
complete. TODO comments document what needs to be done. They will be
addressed in further commits.